change GrowlNotification folder structure
