+++ src/libs/GrowlNotification/GrowlNotificationContainer/index.js
@@ -1,18 +1,22 @@
 import React from 'react';
 import {Animated} from 'react-native';
 import PropTypes from 'prop-types';
-import styles from '../../styles/styles';
+import styles from '../../../styles/styles';
+import withWindowDimensions, {windowDimensionsPropTypes} from '../../../components/withWindowDimensions';
 
 const propTypes = {
     children: PropTypes.node.isRequired,
     translateY: PropTypes.instanceOf(Animated.Value).isRequired,
+    ...windowDimensionsPropTypes,
 };
 
-const GrowlNotificationContainer = ({children, translateY}) => (
+const GrowlNotificationContainer = ({children, translateY, isSmallScreenWidth}) => (
     <Animated.View
         style={[
             styles.growlNotificationContainer,
+            styles.growlNotificationDesktopContainer,
             styles.growlNotificationTranslateY(translateY),
+            isSmallScreenWidth && styles.mwn,
         ]}
     >
         {children}
@@ -21,4 +25,4 @@ const GrowlNotificationContainer = ({children, translateY}) => (
 
 GrowlNotificationContainer.propTypes = propTypes;
 
-export default GrowlNotificationContainer;
+export default withWindowDimensions(GrowlNotificationContainer);
