+++ src/libs/GrowlNotification/GrowlNotificationContainer/index.native.js
@@ -1,22 +1,18 @@
 import React from 'react';
 import {Animated} from 'react-native';
 import PropTypes from 'prop-types';
-import styles from '../../styles/styles';
-import withWindowDimensions, {windowDimensionsPropTypes} from '../../components/withWindowDimensions';
+import styles from '../../../styles/styles';
 
 const propTypes = {
     children: PropTypes.node.isRequired,
     translateY: PropTypes.instanceOf(Animated.Value).isRequired,
-    ...windowDimensionsPropTypes,
 };
 
-const GrowlNotificationContainer = ({children, translateY, isSmallScreenWidth}) => (
+const GrowlNotificationContainer = ({children, translateY}) => (
     <Animated.View
         style={[
             styles.growlNotificationContainer,
-            styles.growlNotificationDesktopContainer,
             styles.growlNotificationTranslateY(translateY),
-            isSmallScreenWidth && styles.mwn,
         ]}
     >
         {children}
@@ -25,4 +21,4 @@ const GrowlNotificationContainer = ({children, translateY, isSmallScreenWidth}) 
 
 GrowlNotificationContainer.propTypes = propTypes;
 
-export default withWindowDimensions(GrowlNotificationContainer);
+export default GrowlNotificationContainer;
